이벤트 목록, 디테일, 쿠폰 수령 생성

+++ ex1/src/main/java/com/example/demo/model/Event.java
@@ -4,13 +4,13 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Getter
-	@Setter
-	public class Event {
-		private Integer event_code ;
-		private String event_title ;
-		private String event_content;
-		private Integer coupon_id;
-		private String event_start;
-		private String event_end;
-	}
+@Setter
+public class Event {
+	private Long event_code;
+	private String event_title;
+	private String event_content;
+	private Integer coupon_id;
+	private String event_start;
+	private String event_end;
+}
